+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# 4.0.1
+* Made animated GIF decoding more reliable
+  * This should fix broken textures #135
+* The maximum size for image downloads is now 10MB by default
+
 # 4.0.0
 * Implemented Texture Resolvers
   * Added support for GIFs #131